requirements.txt
streamlit
absl-py
torch
torchvision
